// src/utils/error-signatures.generated.ts
export const ERROR_METADATA: Record<KnownErrorSignature, ErrorMetadata> = {
  '0x6697b232': {
    name: 'AccessControlBadConfirmation',
    inputs: [],
    source: 'AUT_Roles_v2',
  },
// ...
} as const